Refactor Blockchain Service
