+++ spec/services/blockchain_service/ethereum_spec.rb
@@ -1,7 +1,7 @@
 # encoding: UTF-8
 # frozen_string_literal: true
 
-describe BlockchainService do
+describe BlockchainService::Ethereum do
 
   around do |example|
     WebMock.disable_net_connect!
@@ -9,11 +9,11 @@ describe BlockchainService do
     WebMock.allow_net_connect!
   end
 
-  describe 'BlockAPI::Ethereum' do
+  describe 'Client::Ethereum' do
     let(:block_data) do
       Rails.root.join('spec', 'resources', block_file_name)
-        .yield_self { |file_path| File.open(file_path) }
-        .yield_self { |file| JSON.load(file) }
+          .yield_self { |file_path| File.open(file_path) }
+          .yield_self { |file| JSON.load(file) }
     end
 
     let(:start_block)   { block_data.first['result']['number'].hex }
@@ -21,10 +21,10 @@ describe BlockchainService do
 
     let(:blockchain) do
       Blockchain.find_by_key('eth-rinkeby')
-        .tap { |b| b.update(height: start_block)}
+          .tap { |b| b.update(height: start_block)}
     end
 
-    let(:client) { BlockAPI[blockchain.key] }
+    let(:client) { Client[blockchain.key] }
 
     def request_body(block_number, index)
       { jsonrpc: '2.0',
@@ -41,11 +41,11 @@ describe BlockchainService do
       # Use rinkeby.etherscan.io to fetch transactions data.
       let(:expected_deposits) do
         [
-          {
-            amount:   '0xde0b6b3a7640000'.hex.to_d / currency.base_factor,
-            address:  '0xe3cb6897d83691a8eb8458140a1941ce1d6e6daa',
-            txid:     '0xb60e22c6eed3dc8cd7bc5c7e38c50aa355c55debddbff5c1c4837b995b8ee96d'
-          }
+            {
+                amount:   '0xde0b6b3a7640000'.hex.to_d / currency.base_factor,
+                address:  '0xe3cb6897d83691a8eb8458140a1941ce1d6e6daa',
+                txid:     '0xb60e22c6eed3dc8cd7bc5c7e38c50aa355c55debddbff5c1c4837b995b8ee96d'
+            }
         ]
       end
 
@@ -60,11 +60,11 @@ describe BlockchainService do
         client.class.any_instance.stubs(:latest_block_number).returns(latest_block)
         block_data.each_with_index do |blk, index|
           stub_request(:post, client.endpoint)
-            .with(body: request_body(blk['result']['number'],index))
-            .to_return(body: blk.to_json)
+              .with(body: request_body(blk['result']['number'],index))
+              .to_return(body: blk.to_json)
         end
         # Process blockchain data.
-        BlockchainService.new(blockchain).process_blockchain
+        BlockchainService[blockchain.key].process_blockchain
       end
 
       subject { Deposits::Coin.where(currency: currency) }
@@ -86,7 +86,7 @@ describe BlockchainService do
 
         it 'doesn\'t change deposit' do
           expect(blockchain.height).to eq start_block
-          expect{ BlockchainService.new(blockchain).process_blockchain}.not_to change{subject}
+          expect{ BlockchainService[blockchain.key].process_blockchain}.not_to change{subject}
         end
       end
     end
@@ -98,16 +98,16 @@ describe BlockchainService do
       # Use rinkeby.etherscan.io to fetch transactions data.
       let(:expected_deposits) do
         [
-          {
-            amount:   '0x1e8480'.hex.to_d / currency.base_factor,
-            address:  '0xe3cb6897d83691a8eb8458140a1941ce1d6e6daa',
-            txid:     '0xd5cc0d1d5dd35f4b57572b440fb4ef39a4ab8035657a21692d1871353bfbceea'
-          },
-          {
-            amount:   '0x1e8480'.hex.to_d / currency.base_factor,
-            address:  '0xe3cb6897d83691a8eb8458140a1941ce1d6e6daa',
-            txid:     '0x826555325cec51c4d39b327e563ce3e8ee87e27be5911383f528724a62f0da5d'
-          }
+            {
+                amount:   '0x1e8480'.hex.to_d / currency.base_factor,
+                address:  '0xe3cb6897d83691a8eb8458140a1941ce1d6e6daa',
+                txid:     '0xd5cc0d1d5dd35f4b57572b440fb4ef39a4ab8035657a21692d1871353bfbceea'
+            },
+            {
+                amount:   '0x1e8480'.hex.to_d / currency.base_factor,
+                address:  '0xe3cb6897d83691a8eb8458140a1941ce1d6e6daa',
+                txid:     '0x826555325cec51c4d39b327e563ce3e8ee87e27be5911383f528724a62f0da5d'
+            }
         ]
       end
 
@@ -126,7 +126,7 @@ describe BlockchainService do
               .to_return(body: blk.to_json)
         end
         # Process blockchain data.
-        BlockchainService.new(blockchain).process_blockchain
+        BlockchainService[blockchain.key].process_blockchain
       end
 
       subject { Deposits::Coin.where(currency: currency) }
@@ -148,7 +148,7 @@ describe BlockchainService do
 
         it 'doesn\'t change deposit' do
           expect(blockchain.height).to eq start_block
-          expect{ BlockchainService.new(blockchain).process_blockchain}.not_to change{subject}
+          expect{ BlockchainService[blockchain.key].process_blockchain}.not_to change{subject}
         end
       end
     end
